refactor(server): merge announce + udpbcast into announcer.py
The two files were small, always imported together, and split a single
UDP subsystem across two modules. Merging them into announcer.py (named
after the main class) also removes the name shadow with
protocol/announce.py. SharedUDPServer was never imported anywhere and
is dropped.
